Dedicated Server was having problems, added some extra fixes to make sure the game gets started for players

+++ Source/ProjectCB/CBWorld/CB_GridGenerator.cpp
@@ -5,8 +5,8 @@ const float ACB_GridGenerator::START_SECONDS = 1.0f;
 const float ACB_GridGenerator::UPDATE_INTERVAL = 7.0f;
 const float ACB_GridGenerator::FALL_TIME = 2.0f;
 
-const size_t ACB_GridGenerator::STAGE_WIDTH = 4;
-const size_t ACB_GridGenerator::STAGE_LENGTH = 6;
+const int ACB_GridGenerator::STAGE_WIDTH = 4;
+const int ACB_GridGenerator::STAGE_LENGTH = 6;
 
 const float ACB_GridGenerator::BOX_SIZE = 250.0f;
 const float ACB_GridGenerator::WIDTH_OFFSET = -ACB_GridGenerator::BOX_SIZE * (ACB_GridGenerator::STAGE_WIDTH - 1);
@@ -17,7 +17,7 @@ const float ACB_GridGenerator::BALL_SPAWN_UPDATE_INTERVAL = ACB_GridGenerator::U
 
 void ACB_GridGenerator::deleteBoxes_Implementation()
 {
-    for (size_t index = 0; index < this->m_numOfFallingBoxes; index++) 
+    for (int index = 0; index < this->m_numOfFallingBoxes; index++) 
     {
         if (this->m_fallingBoxes.IsValidIndex(index) && (m_fallingBoxes[index]))
         {
@@ -32,7 +32,7 @@ void ACB_GridGenerator::deleteBoxes_Implementation()
     this->m_numOfFallingBoxes = 0;
 }
 
-ACB_GridBox* ACB_GridGenerator::spawnBox(size_t lengthPos, size_t widthPos)
+ACB_GridBox* ACB_GridGenerator::spawnBox(int lengthPos, int widthPos)
 {
     FActorSpawnParameters spawnParams;
 
@@ -61,19 +61,19 @@ void ACB_GridGenerator::spawnBalls_Implementation()
 {
     if (this->m_numOfBoxes > 0)
     {
-        size_t boxPos = FMath::RandRange(0, this->m_numOfBoxes - 1);
+        int boxPos = FMath::RandRange(0, this->m_numOfBoxes - 1);
 
-        size_t boxID = this->m_deletableBoxes[boxPos];
+        int boxID = this->m_deletableBoxes[boxPos];
 
-        size_t length1 = boxID / ACB_GridGenerator::STAGE_WIDTH;
-        size_t width1 = boxID - (length1 * ACB_GridGenerator::STAGE_WIDTH);
+        int length1 = boxID / ACB_GridGenerator::STAGE_WIDTH;
+        int width1 = boxID - (length1 * ACB_GridGenerator::STAGE_WIDTH);
 
         spawnBall(length1, width1);
 
         if (length1 != (ACB_GridGenerator::STAGE_LENGTH / 2))
         {
-            size_t length2 = ACB_GridGenerator::STAGE_LENGTH - length1 - 1;
-            size_t width2 = ACB_GridGenerator::STAGE_WIDTH - width1 - 1;
+            int length2 = ACB_GridGenerator::STAGE_LENGTH - length1 - 1;
+            int width2 = ACB_GridGenerator::STAGE_WIDTH - width1 - 1;
 
             spawnBall(length2, width2);
         }
@@ -86,11 +86,11 @@ void ACB_GridGenerator::spawnBalls_Implementation()
 
 void ACB_GridGenerator::generateGrid_Implementation()
 {
-    for (size_t length = 0; length < ACB_GridGenerator::STAGE_LENGTH; length++)
+    for (int length = 0; length < ACB_GridGenerator::STAGE_LENGTH; length++)
     {
         TArray<ACB_GridBox*> boxArray;
 
-        for (size_t width = 0; width < ACB_GridGenerator::STAGE_WIDTH; width++)
+        for (int width = 0; width < ACB_GridGenerator::STAGE_WIDTH; width++)
         {
             if (length < ((ACB_GridGenerator::STAGE_LENGTH + 1) / 2))
             {
@@ -111,12 +111,12 @@ void ACB_GridGenerator::updateGrid_Implementation()
     {
         this->m_numOfBoxes--;
 
-        size_t boxPos = FMath::RandRange(0, this->m_numOfBoxes);
+        int boxPos = FMath::RandRange(0, this->m_numOfBoxes);
 
-        size_t boxID = this->m_deletableBoxes[boxPos];
+        int boxID = this->m_deletableBoxes[boxPos];
 
-        size_t length = boxID / ACB_GridGenerator::STAGE_WIDTH;
-        size_t width = boxID - (length * ACB_GridGenerator::STAGE_WIDTH);
+        int length = boxID / ACB_GridGenerator::STAGE_WIDTH;
+        int width = boxID - (length * ACB_GridGenerator::STAGE_WIDTH);
 
         ACB_GridBox* box1 = this->m_grid[length][width];
         box1->startFall();
@@ -209,4 +209,6 @@ void ACB_GridGenerator::GetLifetimeReplicatedProps(TArray<FLifetimeProperty>& Ou
 
     DOREPLIFETIME(ACB_GridGenerator, m_ballHandle);
 }
+
+
